content/pyproject.toml
[project]
name = "hubris-wiki"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-frontmatter (>=1.1.0,<2.0.0)",
    "wikitextparser (>=0.56.4,<0.57.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
